feat(PlaylistPage): move SongItem component into inlined GenericSongItem
